@@ -1,4 +1,5 @@
-from typing import Sequence, TypeVar, Union
+from collections.abc import Sequence
+from typing import TypeVar, Union
 
 import numpy as np
 import numpy.typing as npt
