@@ -1,5 +1,7 @@
 # pylint: disable=missing-class-docstring,missing-function-docstring
-from typing import Generic, Iterator, OrderedDict, Protocol, Tuple, Type, TypeVar
+from collections import OrderedDict
+from collections.abc import Iterator
+from typing import Generic, Protocol, TypeVar
 
 import torch
 from torch import nn
@@ -15,13 +17,13 @@ class ConfigurableModule(Protocol, Generic[C_co]):
     def config(self) -> C_co: ...
 
     @classmethod
-    def load(cls: Type[M], path: PathType) -> M: ...
+    def load(cls: type[M], path: PathType) -> M: ...
 
     def save(self, path: PathType, compile_model: bool = False) -> None: ...
 
     def save_config(self, path: PathType) -> None: ...
 
-    def named_children(self) -> Iterator[Tuple[str, nn.Module]]: ...
+    def named_children(self) -> Iterator[tuple[str, nn.Module]]: ...
 
     def state_dict(self) -> OrderedDict[str, torch.Tensor]: ...
 
