@@ -1,5 +1,4 @@
 # pylint: disable=missing-function-docstring
-from typing import Tuple
 
 import torch
 
@@ -9,7 +8,7 @@ from torchgmm.bayes.gmm import GaussianMixtureModel, GaussianMixtureModelConfig
 
 def sample_gmm(
     num_datapoints: int, num_features: int, num_components: int, covariance_type: CovarianceType
-) -> Tuple[torch.Tensor, torch.Tensor]:
+) -> tuple[torch.Tensor, torch.Tensor]:
     config = GaussianMixtureModelConfig(num_components, num_features, covariance_type)
     model = GaussianMixtureModel(config)
 
